[FEATURE-#89] Refactor AuthenticationFailure and subclasses
AuthenticationFailure is now renamed to AuthenticationError and contains a
message property that contains the error message.
All decendants of the AuthenticationError class now have an Error suffix.

The equals() will return true if both classes are of the same type,
because two errors should be equal if they are of the same type.